fix: Fix an ora2 related test.
I'm not sure why there is an edx-platform test that needs to be updated
for test to ora2 but there is and because of recent changes to ora2, we
needed to update our expected OLX for the platform test.  Making those
changes so we can get the test passing.

I started a conversation to learn more here:
https://github.com/openedx/edx-platform/pull/34145/files#r1598544136